refactor(signal): move Option and config to separate config.go file
- Extract Option interface and config struct from server.go to config.go
- Move optionFunc type and its apply method to config.go
- Transfer WithHandlers and WithRecovery functions to config.go
- Clean up server.go by removing configuration-related code
- Keep server.go focused on Server creation and core logic only
